Extract SAML client to a service
