+++ wazo_auth/plugins/http/saml/http.py
@@ -2,24 +2,12 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 
-import ast
 import logging
 
 from flask import Response, request
-from saml2 import (
-    BINDING_HTTP_POST,
-    BINDING_HTTP_REDIRECT,
-    element_to_extension_element,
-    xmldsig,
-)
-from saml2.client import Saml2Client
-from saml2.config import Config as SAMLConfig
-from saml2.extension.pefim import SPCertEnc
 from saml2.httputil import ServiceError
 from saml2.response import VerificationError
-from saml2.s_utils import UnknownPrincipal, UnsupportedBinding, rndstr
-from saml2.saml import NAMEID_FORMAT_PERSISTENT
-from saml2.samlp import Extensions
+from saml2.s_utils import UnknownPrincipal, UnsupportedBinding
 from saml2.sigver import SignatureError
 
 from wazo_auth import http
@@ -29,52 +17,22 @@ logger = logging.getLogger(__name__)
 
 class SAMLACS(http.ErrorCatchingResource):
     def __init__(
-        self, token_service, user_service, auth_service, config, backend_proxy
+        self,
+        token_service,
+        user_service,
+        auth_service,
+        saml_service,
+        config,
+        backend_proxy,
     ):
         self._token_service = token_service
         self._user_service = user_service
         self._auth_service = auth_service
+        self._saml_service = saml_service
         self._config = (
             config  # Can be used to access to content of the configuration files
         )
         self._backend_proxy = backend_proxy
-        if 'saml' in self._config:
-            try:
-                self._saml_config = SAMLConfig()
-                if isinstance(
-                    self._config['saml']['service']['sp']['endpoints'][
-                        'assertion_consumer_service'
-                    ][0],
-                    str,
-                ):
-                    self._saml_config.load(self._updateConfig(self._config['saml']))
-                else:
-                    self._saml_config.load(self._config['saml'])
-                self._saml_client = Saml2Client(config=self._saml_config)
-            except Exception as inst:
-                logger.error('Error during SAML client init')
-                logger.exception(inst)
-        else:
-            logger.warn(
-                'SAML config is missing, won\'t be able to provide SAML related services'
-            )
-
-    def _extractTuplesFromListOfStrings(self, ls):
-        e = [tuple(i.removeprefix('(').removesuffix(')').split(',')) for i in ls]
-        return [
-            (ast.literal_eval(i[0].strip()), ast.literal_eval(i[1].strip())) for i in e
-        ]
-
-    def _updateConfig(self, config):
-        acs = self._extractTuplesFromListOfStrings(
-            config['service']['sp']['endpoints']['assertion_consumer_service']
-        )
-        slo = self._extractTuplesFromListOfStrings(
-            config['service']['sp']['endpoints']['single_logout_service']
-        )
-        u_config = {'assertion_consumer_service': acs, 'single_logout_service': slo}
-        config['service']['sp']['endpoints'].update(u_config)
-        return config
 
     def post(self):
         # Do you SAML validation here
@@ -88,19 +46,8 @@ class SAMLACS(http.ErrorCatchingResource):
         # }
 
         try:
-            conv_info = {
-                "remote_addr": request.remote_addr,
-                "request_uri": request.url,
-                "entity_id": self._saml_client.config.entityid,
-                "endpoints": self._saml_client.config.getattr("endpoints", "sp"),
-            }
-
-            self.response = self._saml_client.parse_authn_request_response(
-                request.form['SAMLResponse'],
-                BINDING_HTTP_POST,
-                None,
-                None,
-                conv_info=conv_info,
+            self.response = self._saml_service.processAuthResponse(
+                request.url, request.remote_addr, request.form
             )
         except UnknownPrincipal as excp:
             logger.error("UnknownPrincipal: %s", excp)
@@ -111,12 +58,15 @@ class SAMLACS(http.ErrorCatchingResource):
             resp = ServiceError(f"UnsupportedBinding: {excp}")
             return resp(self.environ, self.start_response)
         except VerificationError as err:
+            logger.warn("Verification error: %s", err)
             resp = ServiceError(f"Verification error: {err}")
             return resp(self.environ, self.start_response)
         except SignatureError as err:
+            logger.warn("Signature error: %s", err)
             resp = ServiceError(f"Signature error: {err}")
             return resp(self.environ, self.start_response)
         except Exception as err:
+            logger.error("SAML unexpected error: %s" % err)
             resp = ServiceError(f"Other error: {err}")
             return resp(self.environ, self.start_response)
 
@@ -140,107 +90,28 @@ class SAMLACS(http.ErrorCatchingResource):
 
 class SAMLSSO(http.ErrorCatchingResource):
     def __init__(
-        self, token_service, user_service, auth_service, config, wazo_user_backend
+        self,
+        token_service,
+        user_service,
+        auth_service,
+        saml_service,
+        config,
+        wazo_user_backend,
     ):
         self._token_service = token_service
         self._user_service = user_service
         self._auth_service = auth_service
+        self._saml_service = saml_service
         self._config = config
         self._backend = wazo_user_backend
-        if 'saml' in self._config:
-            try:
-                self._saml_config = SAMLConfig()
-                if isinstance(
-                    self._config['saml']['service']['sp']['endpoints'][
-                        'assertion_consumer_service'
-                    ][0],
-                    str,
-                ):
-                    self._saml_config.load(self._updateConfig(self._config['saml']))
-                else:
-                    self._saml_config.load(self._config['saml'])
-                self._saml_client = Saml2Client(config=self._saml_config)
-            except Exception as inst:
-                logger.error('Error during SAML client init')
-                logger.exception(inst)
-        else:
-            logger.warn(
-                'SAML config is missing, won\'t be able to provide SAML related services'
-            )
-
-    def _extractTuplesFromListOfStrings(self, ls):
-        e = [tuple(i.removeprefix('(').removesuffix(')').split(',')) for i in ls]
-        return [
-            (ast.literal_eval(i[0].strip()), ast.literal_eval(i[1].strip())) for i in e
-        ]
-
-    def _updateConfig(self, config):
-        acs = self._extractTuplesFromListOfStrings(
-            config['service']['sp']['endpoints']['assertion_consumer_service']
-        )
-        slo = self._extractTuplesFromListOfStrings(
-            config['service']['sp']['endpoints']['single_logout_service']
-        )
-        u_config = {'assertion_consumer_service': acs, 'single_logout_service': slo}
-        config['service']['sp']['endpoints'].update(u_config)
-        return config
-
-    def _getSamlRequest(self, saml_client, saml_config):
-        idps = saml_client.metadata.identity_providers()
-
-        entity_id = idps[0]
-
-        _binding, destination = saml_client.pick_binding(
-            "single_sign_on_service",
-            [BINDING_HTTP_REDIRECT],
-            "idpsso",
-            entity_id=entity_id,
-        )
-        logger.debug("binding: %s, destination: %s", _binding, destination)
-        acs = saml_client.config.getattr("endpoints", "sp")[
-            "assertion_consumer_service"
-        ]
-        _, return_binding = acs[0]
-
-        extensions = None
-        if saml_client.config.generate_cert_func is not None:
-            cert_str, req_key_str = saml_client.config.generate_cert_func()
-            spcertenc = SPCertEnc(
-                x509_data=xmldsig.X509Data(
-                    x509_certificate=xmldsig.X509Certificate(text=cert_str)
-                )
-            )
-            extensions = Extensions(
-                extension_elements=[element_to_extension_element(spcertenc)]
-            )
-
-        req_id, req = saml_client.create_authn_request(
-            destination,
-            binding=return_binding,
-            extensions=extensions,
-            nameid_format=NAMEID_FORMAT_PERSISTENT,
-        )
-        _rstate = rndstr()
-        http_args = saml_client.apply_binding(
-            _binding, f"{req}", destination, relay_state=_rstate, sigalg=""
-        )
-        return http_args
-
-    def get(self):
-        if hasattr(self, '_saml_client'):
-            http_args = self._getSamlRequest(self._saml_client, self._saml_config)
-            return Response(headers=http_args['headers'], status=http_args['status'])
-        else:
-            return Response(
-                status=500,
-                response='SAML configuration missing or SAML client init failed',
-            )
 
     def post(self):
-        if hasattr(self, '_saml_client'):
-            http_args = self._getSamlRequest(self._saml_client, self._saml_config)
+        try:
+            # http_args = self._getSamlRequest(self._saml_client, self._saml_config)
+            http_args = self._saml_service.initFlow()
             return Response(headers=http_args['headers'], status=http_args['status'])
-        else:
+        except Exception as excp:
+            logger.error("Failed to process initial SAML SSO post because of: %s", excp)
             return Response(
                 status=500,
                 response='SAML configuration missing or SAML client init failed',
